+++ src/ts/bin/lib/getConfig.ts
@@ -50,6 +50,7 @@ import {
  * @returns  Complete config instance.
  *
  * @since 0.1.0-alpha
+ * @since 0.1.4-alpha — Experimental support for typescript config files.
  * 
  * @internal
  */
@@ -87,6 +88,7 @@ export async function getConfig(
     /** Index of the path currently behind checked. */
     let i = 0;
 
+    /** Maximum number of times to loop while checking for a config file. */
     const maxInterations = pathsToCheck.length;
 
     const fs = new FileSystem( console );
@@ -111,7 +113,7 @@ export async function getConfig(
             continue;
         }
 
-        const content = ( await import( path ) ).default;
+        const content = ( await import( path + '?t=' + Date.now() ) ).default;
 
         console.vi.debug( { content }, 2 + level );